removing boost::variant, moving try_cast into its own header file, make lambda captures explicit
